+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 * [helpdesk] error in helpdesk web ui
+* [services] fix order update which could delete a service from all orders
 
 ## 2.8.0